Multitasking Bugs fixed, Shell partially Implemented, Major Changes/fixes
A New Shell [partially] has been made, which runs as a process. One of
its thread manages console output. All other tasks, when tries to write
to the console, actually writes to a buffer located in the RAM. The
Shell then copies it to the VGA memory every time its processed. Also,
In the earlier version, there was a serious bug which caused the
Multitasking System to crash eventually. They have been fixed now.
Several new fixes have been made and a New Keyboard Driver along with a
fully functional Shell is on its way.